requirements.txt
future==0.18.2
pyad==0.6.0
python-dotenv==0.21.0
